+++ src/adapt/groAdapterNodeLibrary.ts
@@ -2,6 +2,8 @@ import {printSpawnResult, spawn} from '@feltcoop/felt/util/process.js';
 import {EMPTY_OBJECT} from '@feltcoop/felt/util/object.js';
 import {replaceExtension, stripTrailingSlash} from '@feltcoop/felt/util/path.js';
 import {stripStart} from '@feltcoop/felt/util/string.js';
+import esbuild from 'esbuild';
+import {identity} from '@feltcoop/felt/util/function.js';
 
 import {type Adapter} from './adapt.js';
 import {TaskError} from '../task/task.js';
@@ -20,10 +22,10 @@ import {
 import {NODE_LIBRARY_BUILD_NAME} from '../build/buildConfigDefaults.js';
 import {type BuildName} from '../build/buildConfig.js';
 import {printBuildConfigLabel, toInputFiles} from '../build/buildConfig.js';
-import {runRollup} from '../build/rollup.js';
 import {type PathStats} from '../fs/pathData.js';
 import {type PackageJson} from '../utils/packageJson.js';
 import {type Filesystem} from '../fs/filesystem.js';
+import {toDefaultEsbuildBundleOptions} from '../build/groBuilderEsbuildUtils.js';
 
 const name = '@feltcoop/groAdapterNodeLibrary';
 
@@ -60,6 +62,10 @@ export interface Options {
 	bundle: boolean; // defaults to `false`
 }
 
+export interface AdapterArgs {
+	mapBundleOptions?(options: esbuild.BuildOptions): esbuild.BuildOptions;
+}
+
 export const createAdapter = ({
 	buildName = NODE_LIBRARY_BUILD_NAME,
 	dir = `${DIST_DIRNAME}/${buildName}`,
@@ -67,12 +73,12 @@ export const createAdapter = ({
 	packageJson = 'package.json',
 	pack = true,
 	bundle = false,
-}: Partial<Options> = EMPTY_OBJECT): Adapter => {
+}: Partial<Options> = EMPTY_OBJECT): Adapter<AdapterArgs> => {
 	dir = stripTrailingSlash(dir);
 	return {
 		name,
 		adapt: async ({config, fs, dev, log, args, timings}) => {
-			const {mapInputOptions, mapOutputOptions, mapWatchOptions} = args;
+			const {mapBundleOptions = identity} = args;
 
 			const buildConfig = config.builds.find((b) => b.name === buildName);
 			if (!buildConfig) {
@@ -82,7 +88,7 @@ export const createAdapter = ({
 			const files = toInputFiles(buildConfig.input);
 
 			if (bundle) {
-				const timingToBundleWithRollup = timings.start('bundle with rollup');
+				const timingToBundle = timings.start('bundle with rollup');
 				// TODO use `filters` to select the others..right?
 				if (!files.length) {
 					log.trace('no input files in', printBuildConfigLabel(buildConfig));
@@ -91,17 +97,15 @@ export const createAdapter = ({
 				const input = files.map((sourceId) => toImportId(sourceId, dev, buildConfig.name));
 				const outputDir = dir;
 				log.info('bundling', printBuildConfigLabel(buildConfig), outputDir, files);
-				await runRollup({
-					fs,
-					dev,
-					sourcemap: config.sourcemap,
-					input,
-					outputDir,
-					mapInputOptions,
-					mapOutputOptions,
-					mapWatchOptions,
-				});
-				timingToBundleWithRollup();
+				esbuild.build(
+					mapBundleOptions({
+						...toDefaultEsbuildBundleOptions(dev, config.target, config.sourcemap),
+						bundle: true,
+						entryPoints: input,
+						outfile: outputDir + '/index.js',
+					}),
+				);
+				timingToBundle();
 			}
 
 			const timingToCopyDist = timings.start('copy build to dist');